Revert r14947 as this doesn't solve the problem.
This commit was SVN r14949.

The following SVN revision numbers were found above:
  r14947 --> open-mpi/ompi@5b9fe28e3fb163224a5d21dc0d0fa61ce83f655c

+++ ompi/mca/pml/ob1/pml_ob1_sendreq.h
@@ -66,8 +66,8 @@ OBJ_CLASS_DECLARATION(mca_pml_ob1_send_request_t);
 
 struct mca_pml_ob1_send_range_t {
     ompi_free_list_item_t base;
-    size_t range_send_offset;
-    size_t range_send_length;
+    uint64_t range_send_offset;
+    uint64_t range_send_length;
 };
 typedef struct mca_pml_ob1_send_range_t mca_pml_ob1_send_range_t;
 OBJ_CLASS_DECLARATION(mca_pml_ob1_send_range_t);
